consolidate err in stable_result_to_err_no()
