+++ src/utils/err_map.rs
@@ -1,6 +1,10 @@
 use solana_bpf_loader_program::syscalls::SyscallError;
-use solana_program_runtime::solana_rbpf::error::EbpfError;
-use solana_sdk::instruction::InstructionError;
+use solana_program_runtime::{
+    invoke_context::InvokeContext,
+    solana_rbpf::error::{EbpfError, StableResult},
+    stable_log,
+};
+use solana_sdk::{instruction::InstructionError, pubkey::Pubkey};
 
 // Important!
 // The error mapping in this file should be kept aligned with Firedancer.
@@ -135,5 +139,61 @@ pub fn ebpf_err_to_str(error: &EbpfError) -> String {
         EbpfError::LibcInvocationFailed(_, _, _) => "Libc calling returned error code".to_string(),
         EbpfError::VerifierError(_) => "Verifier error".to_string(),
         _ => error.to_string(),
+    }
+}
+
+pub fn stable_result_to_err_no(
+    program_result: StableResult<u64, EbpfError>,
+    invoke_context: &InvokeContext,
+    program_id: &Pubkey,
+) -> i64 {
+    match program_result {
+        StableResult::Ok(_) => 0,
+        StableResult::Err(ref err) => {
+            // Agave/rust propagates errors with additional data, and eventually BPF Loader
+            // logs an error message that depends on the type of error and contains data:
+            // https://github.com/anza-xyz/agave/blob/v2.0.6/program-runtime/src/invoke_context.rs#L535-L549
+            //
+            // Firedancer has a different behavior, it immediately creates the log
+            // when the syscall fails (to avoid propagating data).
+            // Therefore, to match the results, we need to simulate the extra log.
+            //
+            // In the following code we parse error msg and error num in the same way
+            // as Agave does (and logs with stable_log::program_failure()), i.e. by
+            // distinguishing InstructionError, SyscallError or EbpfError.
+            let logger = invoke_context.get_log_collector();
+            let err_no = if let EbpfError::SyscallError(syscall_error) = err {
+                if let Some(instruction_err) = syscall_error.downcast_ref::<InstructionError>() {
+                    stable_log::program_failure(
+                        &logger,
+                        program_id,
+                        &instr_err_to_str(instruction_err),
+                    );
+                    instr_err_to_num(instruction_err)
+                } else if let Some(syscall_error) = syscall_error.downcast_ref::<SyscallError>() {
+                    stable_log::program_failure(
+                        &logger,
+                        program_id,
+                        &syscall_err_to_str(syscall_error),
+                    );
+                    syscall_err_to_num(syscall_error)
+                } else if let Some(syscall_error) = syscall_error.downcast_ref::<EbpfError>() {
+                    stable_log::program_failure(
+                        &logger,
+                        program_id,
+                        &ebpf_err_to_str(syscall_error),
+                    );
+                    ebpf_err_to_num(syscall_error)
+                } else {
+                    // This should never happen, so we return -1 to highlight an unknown error
+                    stable_log::program_failure(&logger, program_id, &ebpf_err_to_str(err));
+                    -1
+                }
+            } else {
+                stable_log::program_failure(&logger, program_id, &ebpf_err_to_str(err));
+                ebpf_err_to_num(err)
+            };
+            err_no as i64
+        }
     }
 }
